DateRange - refactor
