+++ src/components/DateRange/consts.js
@@ -1,0 +1,6 @@
+import { GU } from '../../style'
+
+export const INPUT_HEIGHT = 5 * GU
+export const INPUT_HEIGHT_WITHOUT_BORDER = INPUT_HEIGHT - 2
+export const START_DATE = 'Start date'
+export const END_DATE = 'End date'
